Move PortChooser to its own module
